+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Permissions check to `getUsers` query and `saveUser`, `removeUser` mutations
+
+### Changed
+
+- `getCostCenterByIdStorefront` query ID argument is now optional, will use user's cost center ID as default
+
+### Fixed
+
+- Made storefront queries private
+
 ## [0.5.0] - 2021-11-09
 
 ### Added